Convert prescribers table to use react-pagination
