+++ src/components/Scripts/ScriptsTable.js
@@ -1,38 +1,13 @@
-import { useState, useEffect } from "react";
 import { Link } from "react-router-dom";
 import { useFormatting } from '../../hooks/useFormatting';
 import { StyledPaginate } from "./Pagination.styled";
+import usePagination from "../../hooks/usePagination";
 
 const ScriptsTable = ({ scripts, scriptsPerPage }) => {
   const { formatDrug } = useFormatting();
 
-  // Initialise empty list of current scripts
-  const [currentScripts, setCurrentScripts] = useState([]);
-
-  // Indicates total number of pages required to display all data
-  const [pageCount, setPageCount] = useState(0);
-
-  // Initialise the slice of items at the beginning (index 0)
-  const [scriptsSliceStart, setScriptsSliceStart] = useState(0);
-
-  // Dynamically adjust current items to be displayed when user changes pages
-  useEffect(() => {
-    // Calculate the end index of the items slice required for this page
-    const scriptsSliceEnd = scriptsSliceStart + scriptsPerPage;
-
-    // Slice the items array using the calculated start and end points. This becomes the currently displayed items
-    setCurrentScripts(scripts.slice(scriptsSliceStart, scriptsSliceEnd));
-
-    // Calculate the total pages required to display all the items (to the nearest whole integer)
-    setPageCount(Math.ceil(scripts.length / scriptsPerPage));
-  }, [scriptsSliceStart, scriptsPerPage, scripts]);
-
-  // Invoke when user click to request another page.
-  const handlePageClick = (event) => {
-    // Adjust the new index to slice the data at for the newly selected page
-    const newDataSliceStart = (event.selected * scriptsPerPage) % scripts.length;
-    setScriptsSliceStart(newDataSliceStart);
-  };
+  // This hook controls both the display of scripts on each page, and the numbered pagination controls
+  const { currentItems: currentScripts, pageCount, handlePageClick } = usePagination(scripts, scriptsPerPage);
 
   return (
     <>
