--- src/components/Scripts/testScripts.js
@@ -1,3255 +1,0 @@
-
-let testScripts = [
-  {
-    "brandOnly": false,
-    "authRxNumber": "00032218",
-    "maxRepeats": "5",
-    "maxQuantity": "1",
-    "brandName": "Xalatan",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "quantity": "1",
-    "age": "",
-    "substitutePermitted": true,
-    "itemCode": "5552F",
-    "scriptID": "00003559",
-    "dosage": "Once nightly OU",
-    "indications": "",
-    "prevAuth": false,
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "authCode": "",
-    "verified": true,
-    "authRequired": false,
-    "date": "2022-06-22",
-    "justification": "",
-    "pbsRx": true
-  },
-  {
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "verified": true,
-    "brandName": "Xalatan",
-    "repeats": "5",
-    "date": "2022-06-22",
-    "quantity": "1",
-    "authRequired": false,
-    "indications": "",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "scriptID": "00003558",
-    "prevAuth": false,
-    "maxQuantity": "1",
-    "authRxNumber": "00032207",
-    "itemCode": "5552F",
-    "pbsRx": true,
-    "authCode": "",
-    "brandOnly": false,
-    "age": "",
-    "justification": "",
-    "dosage": "Once nightly OU"
-  },
-  {
-    "date": "2022-06-22",
-    "brandOnly": false,
-    "substitutePermitted": true,
-    "compounded": false,
-    "prevAuth": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "maxQuantity": "1",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "age": "",
-    "includeBrand": true,
-    "authRequired": false,
-    "justification": "",
-    "dosage": "Once nightly OU",
-    "verified": true,
-    "itemCode": "5552F",
-    "quantity": "1",
-    "authRxNumber": "00032196",
-    "scriptID": "00003557",
-    "brandName": "Xalatan",
-    "indications": "",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "authCode": ""
-  },
-  {
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "authRxNumber": "00032185",
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "age": "",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "date": "2022-06-22",
-    "maxQuantity": "1",
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "authCode": "",
-    "scriptID": "00003556",
-    "repeats": "5",
-    "justification": "",
-    "maxRepeats": "5",
-    "brandName": "Xalatan",
-    "authRequired": false,
-    "pbsRx": true,
-    "indications": "",
-    "quantity": "1",
-    "verified": true,
-    "compounded": false,
-    "substitutePermitted": true
-  },
-  {
-    "date": "2022-06-22",
-    "brandName": "Xalatan",
-    "indications": "",
-    "authRxNumber": "00032174",
-    "authRequired": false,
-    "verified": true,
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "compounded": false,
-    "authCode": "",
-    "maxQuantity": "1",
-    "scriptID": "00003555",
-    "prevAuth": false,
-    "pbsRx": true,
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "quantity": "1",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "justification": "",
-    "age": "",
-    "repeats": "5"
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  },
-  {
-    "brandOnly": false,
-    "authRxNumber": "00032218",
-    "maxRepeats": "5",
-    "maxQuantity": "1",
-    "brandName": "Xalatan",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "quantity": "1",
-    "age": "",
-    "substitutePermitted": true,
-    "itemCode": "5552F",
-    "scriptID": "00003559",
-    "dosage": "Once nightly OU",
-    "indications": "",
-    "prevAuth": false,
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "authCode": "",
-    "verified": true,
-    "authRequired": false,
-    "date": "2022-06-22",
-    "justification": "",
-    "pbsRx": true
-  },
-  {
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "verified": true,
-    "brandName": "Xalatan",
-    "repeats": "5",
-    "date": "2022-06-22",
-    "quantity": "1",
-    "authRequired": false,
-    "indications": "",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "scriptID": "00003558",
-    "prevAuth": false,
-    "maxQuantity": "1",
-    "authRxNumber": "00032207",
-    "itemCode": "5552F",
-    "pbsRx": true,
-    "authCode": "",
-    "brandOnly": false,
-    "age": "",
-    "justification": "",
-    "dosage": "Once nightly OU"
-  },
-  {
-    "date": "2022-06-22",
-    "brandOnly": false,
-    "substitutePermitted": true,
-    "compounded": false,
-    "prevAuth": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "maxQuantity": "1",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "age": "",
-    "includeBrand": true,
-    "authRequired": false,
-    "justification": "",
-    "dosage": "Once nightly OU",
-    "verified": true,
-    "itemCode": "5552F",
-    "quantity": "1",
-    "authRxNumber": "00032196",
-    "scriptID": "00003557",
-    "brandName": "Xalatan",
-    "indications": "",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "authCode": ""
-  },
-  {
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "authRxNumber": "00032185",
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "age": "",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "date": "2022-06-22",
-    "maxQuantity": "1",
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "authCode": "",
-    "scriptID": "00003556",
-    "repeats": "5",
-    "justification": "",
-    "maxRepeats": "5",
-    "brandName": "Xalatan",
-    "authRequired": false,
-    "pbsRx": true,
-    "indications": "",
-    "quantity": "1",
-    "verified": true,
-    "compounded": false,
-    "substitutePermitted": true
-  },
-  {
-    "date": "2022-06-22",
-    "brandName": "Xalatan",
-    "indications": "",
-    "authRxNumber": "00032174",
-    "authRequired": false,
-    "verified": true,
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "compounded": false,
-    "authCode": "",
-    "maxQuantity": "1",
-    "scriptID": "00003555",
-    "prevAuth": false,
-    "pbsRx": true,
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "quantity": "1",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "justification": "",
-    "age": "",
-    "repeats": "5"
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  },
-  {
-    "brandOnly": false,
-    "authRxNumber": "00032218",
-    "maxRepeats": "5",
-    "maxQuantity": "1",
-    "brandName": "Xalatan",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "quantity": "1",
-    "age": "",
-    "substitutePermitted": true,
-    "itemCode": "5552F",
-    "scriptID": "00003559",
-    "dosage": "Once nightly OU",
-    "indications": "",
-    "prevAuth": false,
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "authCode": "",
-    "verified": true,
-    "authRequired": false,
-    "date": "2022-06-22",
-    "justification": "",
-    "pbsRx": true
-  },
-  {
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "verified": true,
-    "brandName": "Xalatan",
-    "repeats": "5",
-    "date": "2022-06-22",
-    "quantity": "1",
-    "authRequired": false,
-    "indications": "",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "scriptID": "00003558",
-    "prevAuth": false,
-    "maxQuantity": "1",
-    "authRxNumber": "00032207",
-    "itemCode": "5552F",
-    "pbsRx": true,
-    "authCode": "",
-    "brandOnly": false,
-    "age": "",
-    "justification": "",
-    "dosage": "Once nightly OU"
-  },
-  {
-    "date": "2022-06-22",
-    "brandOnly": false,
-    "substitutePermitted": true,
-    "compounded": false,
-    "prevAuth": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "maxQuantity": "1",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "age": "",
-    "includeBrand": true,
-    "authRequired": false,
-    "justification": "",
-    "dosage": "Once nightly OU",
-    "verified": true,
-    "itemCode": "5552F",
-    "quantity": "1",
-    "authRxNumber": "00032196",
-    "scriptID": "00003557",
-    "brandName": "Xalatan",
-    "indications": "",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "authCode": ""
-  },
-  {
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "authRxNumber": "00032185",
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "age": "",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "date": "2022-06-22",
-    "maxQuantity": "1",
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "authCode": "",
-    "scriptID": "00003556",
-    "repeats": "5",
-    "justification": "",
-    "maxRepeats": "5",
-    "brandName": "Xalatan",
-    "authRequired": false,
-    "pbsRx": true,
-    "indications": "",
-    "quantity": "1",
-    "verified": true,
-    "compounded": false,
-    "substitutePermitted": true
-  },
-  {
-    "date": "2022-06-22",
-    "brandName": "Xalatan",
-    "indications": "",
-    "authRxNumber": "00032174",
-    "authRequired": false,
-    "verified": true,
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "compounded": false,
-    "authCode": "",
-    "maxQuantity": "1",
-    "scriptID": "00003555",
-    "prevAuth": false,
-    "pbsRx": true,
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "quantity": "1",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "justification": "",
-    "age": "",
-    "repeats": "5"
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  },
-  {
-    "brandOnly": false,
-    "authRxNumber": "00032218",
-    "maxRepeats": "5",
-    "maxQuantity": "1",
-    "brandName": "Xalatan",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "quantity": "1",
-    "age": "",
-    "substitutePermitted": true,
-    "itemCode": "5552F",
-    "scriptID": "00003559",
-    "dosage": "Once nightly OU",
-    "indications": "",
-    "prevAuth": false,
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "authCode": "",
-    "verified": true,
-    "authRequired": false,
-    "date": "2022-06-22",
-    "justification": "",
-    "pbsRx": true
-  },
-  {
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "verified": true,
-    "brandName": "Xalatan",
-    "repeats": "5",
-    "date": "2022-06-22",
-    "quantity": "1",
-    "authRequired": false,
-    "indications": "",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "scriptID": "00003558",
-    "prevAuth": false,
-    "maxQuantity": "1",
-    "authRxNumber": "00032207",
-    "itemCode": "5552F",
-    "pbsRx": true,
-    "authCode": "",
-    "brandOnly": false,
-    "age": "",
-    "justification": "",
-    "dosage": "Once nightly OU"
-  },
-  {
-    "date": "2022-06-22",
-    "brandOnly": false,
-    "substitutePermitted": true,
-    "compounded": false,
-    "prevAuth": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "maxQuantity": "1",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "age": "",
-    "includeBrand": true,
-    "authRequired": false,
-    "justification": "",
-    "dosage": "Once nightly OU",
-    "verified": true,
-    "itemCode": "5552F",
-    "quantity": "1",
-    "authRxNumber": "00032196",
-    "scriptID": "00003557",
-    "brandName": "Xalatan",
-    "indications": "",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "authCode": ""
-  },
-  {
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "authRxNumber": "00032185",
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "age": "",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "date": "2022-06-22",
-    "maxQuantity": "1",
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "authCode": "",
-    "scriptID": "00003556",
-    "repeats": "5",
-    "justification": "",
-    "maxRepeats": "5",
-    "brandName": "Xalatan",
-    "authRequired": false,
-    "pbsRx": true,
-    "indications": "",
-    "quantity": "1",
-    "verified": true,
-    "compounded": false,
-    "substitutePermitted": true
-  },
-  {
-    "date": "2022-06-22",
-    "brandName": "Xalatan",
-    "indications": "",
-    "authRxNumber": "00032174",
-    "authRequired": false,
-    "verified": true,
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "compounded": false,
-    "authCode": "",
-    "maxQuantity": "1",
-    "scriptID": "00003555",
-    "prevAuth": false,
-    "pbsRx": true,
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "quantity": "1",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "justification": "",
-    "age": "",
-    "repeats": "5"
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  },
-  {
-    "brandOnly": false,
-    "authRxNumber": "00032218",
-    "maxRepeats": "5",
-    "maxQuantity": "1",
-    "brandName": "Xalatan",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "quantity": "1",
-    "age": "",
-    "substitutePermitted": true,
-    "itemCode": "5552F",
-    "scriptID": "00003559",
-    "dosage": "Once nightly OU",
-    "indications": "",
-    "prevAuth": false,
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "authCode": "",
-    "verified": true,
-    "authRequired": false,
-    "date": "2022-06-22",
-    "justification": "",
-    "pbsRx": true
-  },
-  {
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "verified": true,
-    "brandName": "Xalatan",
-    "repeats": "5",
-    "date": "2022-06-22",
-    "quantity": "1",
-    "authRequired": false,
-    "indications": "",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "scriptID": "00003558",
-    "prevAuth": false,
-    "maxQuantity": "1",
-    "authRxNumber": "00032207",
-    "itemCode": "5552F",
-    "pbsRx": true,
-    "authCode": "",
-    "brandOnly": false,
-    "age": "",
-    "justification": "",
-    "dosage": "Once nightly OU"
-  },
-  {
-    "date": "2022-06-22",
-    "brandOnly": false,
-    "substitutePermitted": true,
-    "compounded": false,
-    "prevAuth": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "maxQuantity": "1",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "age": "",
-    "includeBrand": true,
-    "authRequired": false,
-    "justification": "",
-    "dosage": "Once nightly OU",
-    "verified": true,
-    "itemCode": "5552F",
-    "quantity": "1",
-    "authRxNumber": "00032196",
-    "scriptID": "00003557",
-    "brandName": "Xalatan",
-    "indications": "",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "authCode": ""
-  },
-  {
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "authRxNumber": "00032185",
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "age": "",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "date": "2022-06-22",
-    "maxQuantity": "1",
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "authCode": "",
-    "scriptID": "00003556",
-    "repeats": "5",
-    "justification": "",
-    "maxRepeats": "5",
-    "brandName": "Xalatan",
-    "authRequired": false,
-    "pbsRx": true,
-    "indications": "",
-    "quantity": "1",
-    "verified": true,
-    "compounded": false,
-    "substitutePermitted": true
-  },
-  {
-    "date": "2022-06-22",
-    "brandName": "Xalatan",
-    "indications": "",
-    "authRxNumber": "00032174",
-    "authRequired": false,
-    "verified": true,
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "compounded": false,
-    "authCode": "",
-    "maxQuantity": "1",
-    "scriptID": "00003555",
-    "prevAuth": false,
-    "pbsRx": true,
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "quantity": "1",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "justification": "",
-    "age": "",
-    "repeats": "5"
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  },
-  {
-    "brandOnly": false,
-    "authRxNumber": "00032218",
-    "maxRepeats": "5",
-    "maxQuantity": "1",
-    "brandName": "Xalatan",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "quantity": "1",
-    "age": "",
-    "substitutePermitted": true,
-    "itemCode": "5552F",
-    "scriptID": "00003559",
-    "dosage": "Once nightly OU",
-    "indications": "",
-    "prevAuth": false,
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "authCode": "",
-    "verified": true,
-    "authRequired": false,
-    "date": "2022-06-22",
-    "justification": "",
-    "pbsRx": true
-  },
-  {
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "verified": true,
-    "brandName": "Xalatan",
-    "repeats": "5",
-    "date": "2022-06-22",
-    "quantity": "1",
-    "authRequired": false,
-    "indications": "",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "scriptID": "00003558",
-    "prevAuth": false,
-    "maxQuantity": "1",
-    "authRxNumber": "00032207",
-    "itemCode": "5552F",
-    "pbsRx": true,
-    "authCode": "",
-    "brandOnly": false,
-    "age": "",
-    "justification": "",
-    "dosage": "Once nightly OU"
-  },
-  {
-    "date": "2022-06-22",
-    "brandOnly": false,
-    "substitutePermitted": true,
-    "compounded": false,
-    "prevAuth": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "maxQuantity": "1",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "age": "",
-    "includeBrand": true,
-    "authRequired": false,
-    "justification": "",
-    "dosage": "Once nightly OU",
-    "verified": true,
-    "itemCode": "5552F",
-    "quantity": "1",
-    "authRxNumber": "00032196",
-    "scriptID": "00003557",
-    "brandName": "Xalatan",
-    "indications": "",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "authCode": ""
-  },
-  {
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "authRxNumber": "00032185",
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "age": "",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "date": "2022-06-22",
-    "maxQuantity": "1",
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "authCode": "",
-    "scriptID": "00003556",
-    "repeats": "5",
-    "justification": "",
-    "maxRepeats": "5",
-    "brandName": "Xalatan",
-    "authRequired": false,
-    "pbsRx": true,
-    "indications": "",
-    "quantity": "1",
-    "verified": true,
-    "compounded": false,
-    "substitutePermitted": true
-  },
-  {
-    "date": "2022-06-22",
-    "brandName": "Xalatan",
-    "indications": "",
-    "authRxNumber": "00032174",
-    "authRequired": false,
-    "verified": true,
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "compounded": false,
-    "authCode": "",
-    "maxQuantity": "1",
-    "scriptID": "00003555",
-    "prevAuth": false,
-    "pbsRx": true,
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "quantity": "1",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "justification": "",
-    "age": "",
-    "repeats": "5"
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  },
-  {
-    "brandOnly": false,
-    "authRxNumber": "00032218",
-    "maxRepeats": "5",
-    "maxQuantity": "1",
-    "brandName": "Xalatan",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "quantity": "1",
-    "age": "",
-    "substitutePermitted": true,
-    "itemCode": "5552F",
-    "scriptID": "00003559",
-    "dosage": "Once nightly OU",
-    "indications": "",
-    "prevAuth": false,
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "authCode": "",
-    "verified": true,
-    "authRequired": false,
-    "date": "2022-06-22",
-    "justification": "",
-    "pbsRx": true
-  },
-  {
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "verified": true,
-    "brandName": "Xalatan",
-    "repeats": "5",
-    "date": "2022-06-22",
-    "quantity": "1",
-    "authRequired": false,
-    "indications": "",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "scriptID": "00003558",
-    "prevAuth": false,
-    "maxQuantity": "1",
-    "authRxNumber": "00032207",
-    "itemCode": "5552F",
-    "pbsRx": true,
-    "authCode": "",
-    "brandOnly": false,
-    "age": "",
-    "justification": "",
-    "dosage": "Once nightly OU"
-  },
-  {
-    "date": "2022-06-22",
-    "brandOnly": false,
-    "substitutePermitted": true,
-    "compounded": false,
-    "prevAuth": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "maxQuantity": "1",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "age": "",
-    "includeBrand": true,
-    "authRequired": false,
-    "justification": "",
-    "dosage": "Once nightly OU",
-    "verified": true,
-    "itemCode": "5552F",
-    "quantity": "1",
-    "authRxNumber": "00032196",
-    "scriptID": "00003557",
-    "brandName": "Xalatan",
-    "indications": "",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "authCode": ""
-  },
-  {
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "authRxNumber": "00032185",
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "age": "",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "date": "2022-06-22",
-    "maxQuantity": "1",
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "authCode": "",
-    "scriptID": "00003556",
-    "repeats": "5",
-    "justification": "",
-    "maxRepeats": "5",
-    "brandName": "Xalatan",
-    "authRequired": false,
-    "pbsRx": true,
-    "indications": "",
-    "quantity": "1",
-    "verified": true,
-    "compounded": false,
-    "substitutePermitted": true
-  },
-  {
-    "date": "2022-06-22",
-    "brandName": "Xalatan",
-    "indications": "",
-    "authRxNumber": "00032174",
-    "authRequired": false,
-    "verified": true,
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "compounded": false,
-    "authCode": "",
-    "maxQuantity": "1",
-    "scriptID": "00003555",
-    "prevAuth": false,
-    "pbsRx": true,
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "quantity": "1",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "justification": "",
-    "age": "",
-    "repeats": "5"
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  },
-  {
-    "brandOnly": false,
-    "authRxNumber": "00032218",
-    "maxRepeats": "5",
-    "maxQuantity": "1",
-    "brandName": "Xalatan",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "quantity": "1",
-    "age": "",
-    "substitutePermitted": true,
-    "itemCode": "5552F",
-    "scriptID": "00003559",
-    "dosage": "Once nightly OU",
-    "indications": "",
-    "prevAuth": false,
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "authCode": "",
-    "verified": true,
-    "authRequired": false,
-    "date": "2022-06-22",
-    "justification": "",
-    "pbsRx": true
-  },
-  {
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "verified": true,
-    "brandName": "Xalatan",
-    "repeats": "5",
-    "date": "2022-06-22",
-    "quantity": "1",
-    "authRequired": false,
-    "indications": "",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "scriptID": "00003558",
-    "prevAuth": false,
-    "maxQuantity": "1",
-    "authRxNumber": "00032207",
-    "itemCode": "5552F",
-    "pbsRx": true,
-    "authCode": "",
-    "brandOnly": false,
-    "age": "",
-    "justification": "",
-    "dosage": "Once nightly OU"
-  },
-  {
-    "date": "2022-06-22",
-    "brandOnly": false,
-    "substitutePermitted": true,
-    "compounded": false,
-    "prevAuth": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "maxQuantity": "1",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "age": "",
-    "includeBrand": true,
-    "authRequired": false,
-    "justification": "",
-    "dosage": "Once nightly OU",
-    "verified": true,
-    "itemCode": "5552F",
-    "quantity": "1",
-    "authRxNumber": "00032196",
-    "scriptID": "00003557",
-    "brandName": "Xalatan",
-    "indications": "",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "authCode": ""
-  },
-  {
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "authRxNumber": "00032185",
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "age": "",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "date": "2022-06-22",
-    "maxQuantity": "1",
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "authCode": "",
-    "scriptID": "00003556",
-    "repeats": "5",
-    "justification": "",
-    "maxRepeats": "5",
-    "brandName": "Xalatan",
-    "authRequired": false,
-    "pbsRx": true,
-    "indications": "",
-    "quantity": "1",
-    "verified": true,
-    "compounded": false,
-    "substitutePermitted": true
-  },
-  {
-    "date": "2022-06-22",
-    "brandName": "Xalatan",
-    "indications": "",
-    "authRxNumber": "00032174",
-    "authRequired": false,
-    "verified": true,
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "compounded": false,
-    "authCode": "",
-    "maxQuantity": "1",
-    "scriptID": "00003555",
-    "prevAuth": false,
-    "pbsRx": true,
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "quantity": "1",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "justification": "",
-    "age": "",
-    "repeats": "5"
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  },
-  {
-    "brandOnly": false,
-    "authRxNumber": "00032218",
-    "maxRepeats": "5",
-    "maxQuantity": "1",
-    "brandName": "Xalatan",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "quantity": "1",
-    "age": "",
-    "substitutePermitted": true,
-    "itemCode": "5552F",
-    "scriptID": "00003559",
-    "dosage": "Once nightly OU",
-    "indications": "",
-    "prevAuth": false,
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "authCode": "",
-    "verified": true,
-    "authRequired": false,
-    "date": "2022-06-22",
-    "justification": "",
-    "pbsRx": true
-  },
-  {
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "verified": true,
-    "brandName": "Xalatan",
-    "repeats": "5",
-    "date": "2022-06-22",
-    "quantity": "1",
-    "authRequired": false,
-    "indications": "",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "scriptID": "00003558",
-    "prevAuth": false,
-    "maxQuantity": "1",
-    "authRxNumber": "00032207",
-    "itemCode": "5552F",
-    "pbsRx": true,
-    "authCode": "",
-    "brandOnly": false,
-    "age": "",
-    "justification": "",
-    "dosage": "Once nightly OU"
-  },
-  {
-    "date": "2022-06-22",
-    "brandOnly": false,
-    "substitutePermitted": true,
-    "compounded": false,
-    "prevAuth": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "maxQuantity": "1",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "age": "",
-    "includeBrand": true,
-    "authRequired": false,
-    "justification": "",
-    "dosage": "Once nightly OU",
-    "verified": true,
-    "itemCode": "5552F",
-    "quantity": "1",
-    "authRxNumber": "00032196",
-    "scriptID": "00003557",
-    "brandName": "Xalatan",
-    "indications": "",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "authCode": ""
-  },
-  {
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "authRxNumber": "00032185",
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "age": "",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "date": "2022-06-22",
-    "maxQuantity": "1",
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "authCode": "",
-    "scriptID": "00003556",
-    "repeats": "5",
-    "justification": "",
-    "maxRepeats": "5",
-    "brandName": "Xalatan",
-    "authRequired": false,
-    "pbsRx": true,
-    "indications": "",
-    "quantity": "1",
-    "verified": true,
-    "compounded": false,
-    "substitutePermitted": true
-  },
-  {
-    "date": "2022-06-22",
-    "brandName": "Xalatan",
-    "indications": "",
-    "authRxNumber": "00032174",
-    "authRequired": false,
-    "verified": true,
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "compounded": false,
-    "authCode": "",
-    "maxQuantity": "1",
-    "scriptID": "00003555",
-    "prevAuth": false,
-    "pbsRx": true,
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "quantity": "1",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "justification": "",
-    "age": "",
-    "repeats": "5"
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  },
-  {
-    "brandOnly": false,
-    "authRxNumber": "00032218",
-    "maxRepeats": "5",
-    "maxQuantity": "1",
-    "brandName": "Xalatan",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "quantity": "1",
-    "age": "",
-    "substitutePermitted": true,
-    "itemCode": "5552F",
-    "scriptID": "00003559",
-    "dosage": "Once nightly OU",
-    "indications": "",
-    "prevAuth": false,
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "authCode": "",
-    "verified": true,
-    "authRequired": false,
-    "date": "2022-06-22",
-    "justification": "",
-    "pbsRx": true
-  },
-  {
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "verified": true,
-    "brandName": "Xalatan",
-    "repeats": "5",
-    "date": "2022-06-22",
-    "quantity": "1",
-    "authRequired": false,
-    "indications": "",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "scriptID": "00003558",
-    "prevAuth": false,
-    "maxQuantity": "1",
-    "authRxNumber": "00032207",
-    "itemCode": "5552F",
-    "pbsRx": true,
-    "authCode": "",
-    "brandOnly": false,
-    "age": "",
-    "justification": "",
-    "dosage": "Once nightly OU"
-  },
-  {
-    "date": "2022-06-22",
-    "brandOnly": false,
-    "substitutePermitted": true,
-    "compounded": false,
-    "prevAuth": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "maxQuantity": "1",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "age": "",
-    "includeBrand": true,
-    "authRequired": false,
-    "justification": "",
-    "dosage": "Once nightly OU",
-    "verified": true,
-    "itemCode": "5552F",
-    "quantity": "1",
-    "authRxNumber": "00032196",
-    "scriptID": "00003557",
-    "brandName": "Xalatan",
-    "indications": "",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "authCode": ""
-  },
-  {
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "authRxNumber": "00032185",
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "age": "",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "date": "2022-06-22",
-    "maxQuantity": "1",
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "authCode": "",
-    "scriptID": "00003556",
-    "repeats": "5",
-    "justification": "",
-    "maxRepeats": "5",
-    "brandName": "Xalatan",
-    "authRequired": false,
-    "pbsRx": true,
-    "indications": "",
-    "quantity": "1",
-    "verified": true,
-    "compounded": false,
-    "substitutePermitted": true
-  },
-  {
-    "date": "2022-06-22",
-    "brandName": "Xalatan",
-    "indications": "",
-    "authRxNumber": "00032174",
-    "authRequired": false,
-    "verified": true,
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "compounded": false,
-    "authCode": "",
-    "maxQuantity": "1",
-    "scriptID": "00003555",
-    "prevAuth": false,
-    "pbsRx": true,
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "quantity": "1",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "justification": "",
-    "age": "",
-    "repeats": "5"
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  },
-  {
-    "brandOnly": false,
-    "authRxNumber": "00032218",
-    "maxRepeats": "5",
-    "maxQuantity": "1",
-    "brandName": "Xalatan",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "quantity": "1",
-    "age": "",
-    "substitutePermitted": true,
-    "itemCode": "5552F",
-    "scriptID": "00003559",
-    "dosage": "Once nightly OU",
-    "indications": "",
-    "prevAuth": false,
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "authCode": "",
-    "verified": true,
-    "authRequired": false,
-    "date": "2022-06-22",
-    "justification": "",
-    "pbsRx": true
-  },
-  {
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "verified": true,
-    "brandName": "Xalatan",
-    "repeats": "5",
-    "date": "2022-06-22",
-    "quantity": "1",
-    "authRequired": false,
-    "indications": "",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "scriptID": "00003558",
-    "prevAuth": false,
-    "maxQuantity": "1",
-    "authRxNumber": "00032207",
-    "itemCode": "5552F",
-    "pbsRx": true,
-    "authCode": "",
-    "brandOnly": false,
-    "age": "",
-    "justification": "",
-    "dosage": "Once nightly OU"
-  },
-  {
-    "date": "2022-06-22",
-    "brandOnly": false,
-    "substitutePermitted": true,
-    "compounded": false,
-    "prevAuth": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "maxQuantity": "1",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "age": "",
-    "includeBrand": true,
-    "authRequired": false,
-    "justification": "",
-    "dosage": "Once nightly OU",
-    "verified": true,
-    "itemCode": "5552F",
-    "quantity": "1",
-    "authRxNumber": "00032196",
-    "scriptID": "00003557",
-    "brandName": "Xalatan",
-    "indications": "",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "authCode": ""
-  },
-  {
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "authRxNumber": "00032185",
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "age": "",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "date": "2022-06-22",
-    "maxQuantity": "1",
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "authCode": "",
-    "scriptID": "00003556",
-    "repeats": "5",
-    "justification": "",
-    "maxRepeats": "5",
-    "brandName": "Xalatan",
-    "authRequired": false,
-    "pbsRx": true,
-    "indications": "",
-    "quantity": "1",
-    "verified": true,
-    "compounded": false,
-    "substitutePermitted": true
-  },
-  {
-    "date": "2022-06-22",
-    "brandName": "Xalatan",
-    "indications": "",
-    "authRxNumber": "00032174",
-    "authRequired": false,
-    "verified": true,
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "compounded": false,
-    "authCode": "",
-    "maxQuantity": "1",
-    "scriptID": "00003555",
-    "prevAuth": false,
-    "pbsRx": true,
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "quantity": "1",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "justification": "",
-    "age": "",
-    "repeats": "5"
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  },
-  {
-    "brandOnly": false,
-    "authRxNumber": "00032218",
-    "maxRepeats": "5",
-    "maxQuantity": "1",
-    "brandName": "Xalatan",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "quantity": "1",
-    "age": "",
-    "substitutePermitted": true,
-    "itemCode": "5552F",
-    "scriptID": "00003559",
-    "dosage": "Once nightly OU",
-    "indications": "",
-    "prevAuth": false,
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "authCode": "",
-    "verified": true,
-    "authRequired": false,
-    "date": "2022-06-22",
-    "justification": "",
-    "pbsRx": true
-  },
-  {
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "verified": true,
-    "brandName": "Xalatan",
-    "repeats": "5",
-    "date": "2022-06-22",
-    "quantity": "1",
-    "authRequired": false,
-    "indications": "",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "scriptID": "00003558",
-    "prevAuth": false,
-    "maxQuantity": "1",
-    "authRxNumber": "00032207",
-    "itemCode": "5552F",
-    "pbsRx": true,
-    "authCode": "",
-    "brandOnly": false,
-    "age": "",
-    "justification": "",
-    "dosage": "Once nightly OU"
-  },
-  {
-    "date": "2022-06-22",
-    "brandOnly": false,
-    "substitutePermitted": true,
-    "compounded": false,
-    "prevAuth": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "maxQuantity": "1",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "age": "",
-    "includeBrand": true,
-    "authRequired": false,
-    "justification": "",
-    "dosage": "Once nightly OU",
-    "verified": true,
-    "itemCode": "5552F",
-    "quantity": "1",
-    "authRxNumber": "00032196",
-    "scriptID": "00003557",
-    "brandName": "Xalatan",
-    "indications": "",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "authCode": ""
-  },
-  {
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "authRxNumber": "00032185",
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "age": "",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "date": "2022-06-22",
-    "maxQuantity": "1",
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "authCode": "",
-    "scriptID": "00003556",
-    "repeats": "5",
-    "justification": "",
-    "maxRepeats": "5",
-    "brandName": "Xalatan",
-    "authRequired": false,
-    "pbsRx": true,
-    "indications": "",
-    "quantity": "1",
-    "verified": true,
-    "compounded": false,
-    "substitutePermitted": true
-  },
-  {
-    "date": "2022-06-22",
-    "brandName": "Xalatan",
-    "indications": "",
-    "authRxNumber": "00032174",
-    "authRequired": false,
-    "verified": true,
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "compounded": false,
-    "authCode": "",
-    "maxQuantity": "1",
-    "scriptID": "00003555",
-    "prevAuth": false,
-    "pbsRx": true,
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "quantity": "1",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "justification": "",
-    "age": "",
-    "repeats": "5"
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  },
-  {
-    "brandOnly": false,
-    "authRxNumber": "00032218",
-    "maxRepeats": "5",
-    "maxQuantity": "1",
-    "brandName": "Xalatan",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "quantity": "1",
-    "age": "",
-    "substitutePermitted": true,
-    "itemCode": "5552F",
-    "scriptID": "00003559",
-    "dosage": "Once nightly OU",
-    "indications": "",
-    "prevAuth": false,
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "authCode": "",
-    "verified": true,
-    "authRequired": false,
-    "date": "2022-06-22",
-    "justification": "",
-    "pbsRx": true
-  },
-  {
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "verified": true,
-    "brandName": "Xalatan",
-    "repeats": "5",
-    "date": "2022-06-22",
-    "quantity": "1",
-    "authRequired": false,
-    "indications": "",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "scriptID": "00003558",
-    "prevAuth": false,
-    "maxQuantity": "1",
-    "authRxNumber": "00032207",
-    "itemCode": "5552F",
-    "pbsRx": true,
-    "authCode": "",
-    "brandOnly": false,
-    "age": "",
-    "justification": "",
-    "dosage": "Once nightly OU"
-  },
-  {
-    "date": "2022-06-22",
-    "brandOnly": false,
-    "substitutePermitted": true,
-    "compounded": false,
-    "prevAuth": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "maxQuantity": "1",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "age": "",
-    "includeBrand": true,
-    "authRequired": false,
-    "justification": "",
-    "dosage": "Once nightly OU",
-    "verified": true,
-    "itemCode": "5552F",
-    "quantity": "1",
-    "authRxNumber": "00032196",
-    "scriptID": "00003557",
-    "brandName": "Xalatan",
-    "indications": "",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "authCode": ""
-  },
-  {
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "authRxNumber": "00032185",
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "age": "",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "date": "2022-06-22",
-    "maxQuantity": "1",
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "authCode": "",
-    "scriptID": "00003556",
-    "repeats": "5",
-    "justification": "",
-    "maxRepeats": "5",
-    "brandName": "Xalatan",
-    "authRequired": false,
-    "pbsRx": true,
-    "indications": "",
-    "quantity": "1",
-    "verified": true,
-    "compounded": false,
-    "substitutePermitted": true
-  },
-  {
-    "date": "2022-06-22",
-    "brandName": "Xalatan",
-    "indications": "",
-    "authRxNumber": "00032174",
-    "authRequired": false,
-    "verified": true,
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "compounded": false,
-    "authCode": "",
-    "maxQuantity": "1",
-    "scriptID": "00003555",
-    "prevAuth": false,
-    "pbsRx": true,
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "quantity": "1",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "justification": "",
-    "age": "",
-    "repeats": "5"
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  },
-  {
-    "brandOnly": false,
-    "authRxNumber": "00032218",
-    "maxRepeats": "5",
-    "maxQuantity": "1",
-    "brandName": "Xalatan",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "quantity": "1",
-    "age": "",
-    "substitutePermitted": true,
-    "itemCode": "5552F",
-    "scriptID": "00003559",
-    "dosage": "Once nightly OU",
-    "indications": "",
-    "prevAuth": false,
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "authCode": "",
-    "verified": true,
-    "authRequired": false,
-    "date": "2022-06-22",
-    "justification": "",
-    "pbsRx": true
-  },
-  {
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "verified": true,
-    "brandName": "Xalatan",
-    "repeats": "5",
-    "date": "2022-06-22",
-    "quantity": "1",
-    "authRequired": false,
-    "indications": "",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "scriptID": "00003558",
-    "prevAuth": false,
-    "maxQuantity": "1",
-    "authRxNumber": "00032207",
-    "itemCode": "5552F",
-    "pbsRx": true,
-    "authCode": "",
-    "brandOnly": false,
-    "age": "",
-    "justification": "",
-    "dosage": "Once nightly OU"
-  },
-  {
-    "date": "2022-06-22",
-    "brandOnly": false,
-    "substitutePermitted": true,
-    "compounded": false,
-    "prevAuth": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "maxQuantity": "1",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "age": "",
-    "includeBrand": true,
-    "authRequired": false,
-    "justification": "",
-    "dosage": "Once nightly OU",
-    "verified": true,
-    "itemCode": "5552F",
-    "quantity": "1",
-    "authRxNumber": "00032196",
-    "scriptID": "00003557",
-    "brandName": "Xalatan",
-    "indications": "",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "authCode": ""
-  },
-  {
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "authRxNumber": "00032185",
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "age": "",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "date": "2022-06-22",
-    "maxQuantity": "1",
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "authCode": "",
-    "scriptID": "00003556",
-    "repeats": "5",
-    "justification": "",
-    "maxRepeats": "5",
-    "brandName": "Xalatan",
-    "authRequired": false,
-    "pbsRx": true,
-    "indications": "",
-    "quantity": "1",
-    "verified": true,
-    "compounded": false,
-    "substitutePermitted": true
-  },
-  {
-    "date": "2022-06-22",
-    "brandName": "Xalatan",
-    "indications": "",
-    "authRxNumber": "00032174",
-    "authRequired": false,
-    "verified": true,
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "compounded": false,
-    "authCode": "",
-    "maxQuantity": "1",
-    "scriptID": "00003555",
-    "prevAuth": false,
-    "pbsRx": true,
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "quantity": "1",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "justification": "",
-    "age": "",
-    "repeats": "5"
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  },
-  {
-    "date": "2022-06-22",
-    "brandOnly": false,
-    "substitutePermitted": true,
-    "compounded": false,
-    "prevAuth": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "maxQuantity": "1",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "age": "",
-    "includeBrand": true,
-    "authRequired": false,
-    "justification": "",
-    "dosage": "Once nightly OU",
-    "verified": true,
-    "itemCode": "5552F",
-    "quantity": "1",
-    "authRxNumber": "00032196",
-    "scriptID": "00003557",
-    "brandName": "Xalatan",
-    "indications": "",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "authCode": ""
-  },
-  {
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "authRxNumber": "00032185",
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "age": "",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "date": "2022-06-22",
-    "maxQuantity": "1",
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "authCode": "",
-    "scriptID": "00003556",
-    "repeats": "5",
-    "justification": "",
-    "maxRepeats": "5",
-    "brandName": "Xalatan",
-    "authRequired": false,
-    "pbsRx": true,
-    "indications": "",
-    "quantity": "1",
-    "verified": true,
-    "compounded": false,
-    "substitutePermitted": true
-  },
-  {
-    "date": "2022-06-22",
-    "brandName": "Xalatan",
-    "indications": "",
-    "authRxNumber": "00032174",
-    "authRequired": false,
-    "verified": true,
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "compounded": false,
-    "authCode": "",
-    "maxQuantity": "1",
-    "scriptID": "00003555",
-    "prevAuth": false,
-    "pbsRx": true,
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "quantity": "1",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "justification": "",
-    "age": "",
-    "repeats": "5"
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  },
-  {
-    "date": "2022-06-22",
-    "brandOnly": false,
-    "substitutePermitted": true,
-    "compounded": false,
-    "prevAuth": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "maxQuantity": "1",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "age": "",
-    "includeBrand": true,
-    "authRequired": false,
-    "justification": "",
-    "dosage": "Once nightly OU",
-    "verified": true,
-    "itemCode": "5552F",
-    "quantity": "1",
-    "authRxNumber": "00032196",
-    "scriptID": "00003557",
-    "brandName": "Xalatan",
-    "indications": "",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "authCode": ""
-  },
-  {
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "authRxNumber": "00032185",
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "age": "",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "date": "2022-06-22",
-    "maxQuantity": "1",
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "authCode": "",
-    "scriptID": "00003556",
-    "repeats": "5",
-    "justification": "",
-    "maxRepeats": "5",
-    "brandName": "Xalatan",
-    "authRequired": false,
-    "pbsRx": true,
-    "indications": "",
-    "quantity": "1",
-    "verified": true,
-    "compounded": false,
-    "substitutePermitted": true
-  },
-  {
-    "date": "2022-06-22",
-    "brandName": "Xalatan",
-    "indications": "",
-    "authRxNumber": "00032174",
-    "authRequired": false,
-    "verified": true,
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "compounded": false,
-    "authCode": "",
-    "maxQuantity": "1",
-    "scriptID": "00003555",
-    "prevAuth": false,
-    "pbsRx": true,
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "quantity": "1",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "justification": "",
-    "age": "",
-    "repeats": "5"
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  },
-  {
-    "brandOnly": false,
-    "authRxNumber": "00032218",
-    "maxRepeats": "5",
-    "maxQuantity": "1",
-    "brandName": "Xalatan",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "quantity": "1",
-    "age": "",
-    "substitutePermitted": true,
-    "itemCode": "5552F",
-    "scriptID": "00003559",
-    "dosage": "Once nightly OU",
-    "indications": "",
-    "prevAuth": false,
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "authCode": "",
-    "verified": true,
-    "authRequired": false,
-    "date": "2022-06-22",
-    "justification": "",
-    "pbsRx": true
-  },
-  {
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "verified": true,
-    "brandName": "Xalatan",
-    "repeats": "5",
-    "date": "2022-06-22",
-    "quantity": "1",
-    "authRequired": false,
-    "indications": "",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "scriptID": "00003558",
-    "prevAuth": false,
-    "maxQuantity": "1",
-    "authRxNumber": "00032207",
-    "itemCode": "5552F",
-    "pbsRx": true,
-    "authCode": "",
-    "brandOnly": false,
-    "age": "",
-    "justification": "",
-    "dosage": "Once nightly OU"
-  },
-  {
-    "date": "2022-06-22",
-    "brandOnly": false,
-    "substitutePermitted": true,
-    "compounded": false,
-    "prevAuth": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "maxQuantity": "1",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "age": "",
-    "includeBrand": true,
-    "authRequired": false,
-    "justification": "",
-    "dosage": "Once nightly OU",
-    "verified": true,
-    "itemCode": "5552F",
-    "quantity": "1",
-    "authRxNumber": "00032196",
-    "scriptID": "00003557",
-    "brandName": "Xalatan",
-    "indications": "",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "authCode": ""
-  },
-  {
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "authRxNumber": "00032185",
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "age": "",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "date": "2022-06-22",
-    "maxQuantity": "1",
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "authCode": "",
-    "scriptID": "00003556",
-    "repeats": "5",
-    "justification": "",
-    "maxRepeats": "5",
-    "brandName": "Xalatan",
-    "authRequired": false,
-    "pbsRx": true,
-    "indications": "",
-    "quantity": "1",
-    "verified": true,
-    "compounded": false,
-    "substitutePermitted": true
-  },
-  {
-    "date": "2022-06-22",
-    "brandName": "Xalatan",
-    "indications": "",
-    "authRxNumber": "00032174",
-    "authRequired": false,
-    "verified": true,
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "compounded": false,
-    "authCode": "",
-    "maxQuantity": "1",
-    "scriptID": "00003555",
-    "prevAuth": false,
-    "pbsRx": true,
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "quantity": "1",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "justification": "",
-    "age": "",
-    "repeats": "5"
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  },
-  {
-    "brandOnly": false,
-    "authRxNumber": "00032218",
-    "maxRepeats": "5",
-    "maxQuantity": "1",
-    "brandName": "Xalatan",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "quantity": "1",
-    "age": "",
-    "substitutePermitted": true,
-    "itemCode": "5552F",
-    "scriptID": "00003559",
-    "dosage": "Once nightly OU",
-    "indications": "",
-    "prevAuth": false,
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "authCode": "",
-    "verified": true,
-    "authRequired": false,
-    "date": "2022-06-22",
-    "justification": "",
-    "pbsRx": true
-  },
-  {
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "verified": true,
-    "brandName": "Xalatan",
-    "repeats": "5",
-    "date": "2022-06-22",
-    "quantity": "1",
-    "authRequired": false,
-    "indications": "",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "scriptID": "00003558",
-    "prevAuth": false,
-    "maxQuantity": "1",
-    "authRxNumber": "00032207",
-    "itemCode": "5552F",
-    "pbsRx": true,
-    "authCode": "",
-    "brandOnly": false,
-    "age": "",
-    "justification": "",
-    "dosage": "Once nightly OU"
-  },
-  {
-    "date": "2022-06-22",
-    "brandOnly": false,
-    "substitutePermitted": true,
-    "compounded": false,
-    "prevAuth": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "maxQuantity": "1",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "age": "",
-    "includeBrand": true,
-    "authRequired": false,
-    "justification": "",
-    "dosage": "Once nightly OU",
-    "verified": true,
-    "itemCode": "5552F",
-    "quantity": "1",
-    "authRxNumber": "00032196",
-    "scriptID": "00003557",
-    "brandName": "Xalatan",
-    "indications": "",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "authCode": ""
-  },
-  {
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "authRxNumber": "00032185",
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "age": "",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "date": "2022-06-22",
-    "maxQuantity": "1",
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "authCode": "",
-    "scriptID": "00003556",
-    "repeats": "5",
-    "justification": "",
-    "maxRepeats": "5",
-    "brandName": "Xalatan",
-    "authRequired": false,
-    "pbsRx": true,
-    "indications": "",
-    "quantity": "1",
-    "verified": true,
-    "compounded": false,
-    "substitutePermitted": true
-  },
-  {
-    "date": "2022-06-22",
-    "brandName": "Xalatan",
-    "indications": "",
-    "authRxNumber": "00032174",
-    "authRequired": false,
-    "verified": true,
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "compounded": false,
-    "authCode": "",
-    "maxQuantity": "1",
-    "scriptID": "00003555",
-    "prevAuth": false,
-    "pbsRx": true,
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "quantity": "1",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "justification": "",
-    "age": "",
-    "repeats": "5"
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  },
-  {
-    "brandOnly": false,
-    "authRxNumber": "00032218",
-    "maxRepeats": "5",
-    "maxQuantity": "1",
-    "brandName": "Xalatan",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "quantity": "1",
-    "age": "",
-    "substitutePermitted": true,
-    "itemCode": "5552F",
-    "scriptID": "00003559",
-    "dosage": "Once nightly OU",
-    "indications": "",
-    "prevAuth": false,
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "authCode": "",
-    "verified": true,
-    "authRequired": false,
-    "date": "2022-06-22",
-    "justification": "",
-    "pbsRx": true
-  },
-  {
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "verified": true,
-    "brandName": "Xalatan",
-    "repeats": "5",
-    "date": "2022-06-22",
-    "quantity": "1",
-    "authRequired": false,
-    "indications": "",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "scriptID": "00003558",
-    "prevAuth": false,
-    "maxQuantity": "1",
-    "authRxNumber": "00032207",
-    "itemCode": "5552F",
-    "pbsRx": true,
-    "authCode": "",
-    "brandOnly": false,
-    "age": "",
-    "justification": "",
-    "dosage": "Once nightly OU"
-  },
-  {
-    "date": "2022-06-22",
-    "brandOnly": false,
-    "substitutePermitted": true,
-    "compounded": false,
-    "prevAuth": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "maxQuantity": "1",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "age": "",
-    "includeBrand": true,
-    "authRequired": false,
-    "justification": "",
-    "dosage": "Once nightly OU",
-    "verified": true,
-    "itemCode": "5552F",
-    "quantity": "1",
-    "authRxNumber": "00032196",
-    "scriptID": "00003557",
-    "brandName": "Xalatan",
-    "indications": "",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "authCode": ""
-  },
-  {
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "authRxNumber": "00032185",
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "age": "",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "date": "2022-06-22",
-    "maxQuantity": "1",
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "authCode": "",
-    "scriptID": "00003556",
-    "repeats": "5",
-    "justification": "",
-    "maxRepeats": "5",
-    "brandName": "Xalatan",
-    "authRequired": false,
-    "pbsRx": true,
-    "indications": "",
-    "quantity": "1",
-    "verified": true,
-    "compounded": false,
-    "substitutePermitted": true
-  },
-  {
-    "date": "2022-06-22",
-    "brandName": "Xalatan",
-    "indications": "",
-    "authRxNumber": "00032174",
-    "authRequired": false,
-    "verified": true,
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "compounded": false,
-    "authCode": "",
-    "maxQuantity": "1",
-    "scriptID": "00003555",
-    "prevAuth": false,
-    "pbsRx": true,
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "quantity": "1",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "justification": "",
-    "age": "",
-    "repeats": "5"
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  },
-  {
-    "brandOnly": false,
-    "authRxNumber": "00032218",
-    "maxRepeats": "5",
-    "maxQuantity": "1",
-    "brandName": "Xalatan",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "quantity": "1",
-    "age": "",
-    "substitutePermitted": true,
-    "itemCode": "5552F",
-    "scriptID": "00003559",
-    "dosage": "Once nightly OU",
-    "indications": "",
-    "prevAuth": false,
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "authCode": "",
-    "verified": true,
-    "authRequired": false,
-    "date": "2022-06-22",
-    "justification": "",
-    "pbsRx": true
-  },
-  {
-    "includeBrand": true,
-    "compounded": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "verified": true,
-    "brandName": "Xalatan",
-    "repeats": "5",
-    "date": "2022-06-22",
-    "quantity": "1",
-    "authRequired": false,
-    "indications": "",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "scriptID": "00003558",
-    "prevAuth": false,
-    "maxQuantity": "1",
-    "authRxNumber": "00032207",
-    "itemCode": "5552F",
-    "pbsRx": true,
-    "authCode": "",
-    "brandOnly": false,
-    "age": "",
-    "justification": "",
-    "dosage": "Once nightly OU"
-  },
-  {
-    "date": "2022-06-22",
-    "brandOnly": false,
-    "substitutePermitted": true,
-    "compounded": false,
-    "prevAuth": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "maxQuantity": "1",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "age": "",
-    "includeBrand": true,
-    "authRequired": false,
-    "justification": "",
-    "dosage": "Once nightly OU",
-    "verified": true,
-    "itemCode": "5552F",
-    "quantity": "1",
-    "authRxNumber": "00032196",
-    "scriptID": "00003557",
-    "brandName": "Xalatan",
-    "indications": "",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "authCode": ""
-  },
-  {
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "authRxNumber": "00032185",
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "age": "",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "date": "2022-06-22",
-    "maxQuantity": "1",
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "authCode": "",
-    "scriptID": "00003556",
-    "repeats": "5",
-    "justification": "",
-    "maxRepeats": "5",
-    "brandName": "Xalatan",
-    "authRequired": false,
-    "pbsRx": true,
-    "indications": "",
-    "quantity": "1",
-    "verified": true,
-    "compounded": false,
-    "substitutePermitted": true
-  },
-  {
-    "date": "2022-06-22",
-    "brandName": "Xalatan",
-    "indications": "",
-    "authRxNumber": "00032174",
-    "authRequired": false,
-    "verified": true,
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "compounded": false,
-    "authCode": "",
-    "maxQuantity": "1",
-    "scriptID": "00003555",
-    "prevAuth": false,
-    "pbsRx": true,
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "quantity": "1",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "justification": "",
-    "age": "",
-    "repeats": "5"
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  },
-  {
-    "date": "2022-06-22",
-    "brandOnly": false,
-    "substitutePermitted": true,
-    "compounded": false,
-    "prevAuth": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "maxQuantity": "1",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "age": "",
-    "includeBrand": true,
-    "authRequired": false,
-    "justification": "",
-    "dosage": "Once nightly OU",
-    "verified": true,
-    "itemCode": "5552F",
-    "quantity": "1",
-    "authRxNumber": "00032196",
-    "scriptID": "00003557",
-    "brandName": "Xalatan",
-    "indications": "",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "authCode": ""
-  },
-  {
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "authRxNumber": "00032185",
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "age": "",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "date": "2022-06-22",
-    "maxQuantity": "1",
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "authCode": "",
-    "scriptID": "00003556",
-    "repeats": "5",
-    "justification": "",
-    "maxRepeats": "5",
-    "brandName": "Xalatan",
-    "authRequired": false,
-    "pbsRx": true,
-    "indications": "",
-    "quantity": "1",
-    "verified": true,
-    "compounded": false,
-    "substitutePermitted": true
-  },
-  {
-    "date": "2022-06-22",
-    "brandName": "Xalatan",
-    "indications": "",
-    "authRxNumber": "00032174",
-    "authRequired": false,
-    "verified": true,
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "compounded": false,
-    "authCode": "",
-    "maxQuantity": "1",
-    "scriptID": "00003555",
-    "prevAuth": false,
-    "pbsRx": true,
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "quantity": "1",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "justification": "",
-    "age": "",
-    "repeats": "5"
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  },
-  {
-    "date": "2022-06-22",
-    "brandOnly": false,
-    "substitutePermitted": true,
-    "compounded": false,
-    "prevAuth": false,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "maxQuantity": "1",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "repeats": "5",
-    "age": "",
-    "includeBrand": true,
-    "authRequired": false,
-    "justification": "",
-    "dosage": "Once nightly OU",
-    "verified": true,
-    "itemCode": "5552F",
-    "quantity": "1",
-    "authRxNumber": "00032196",
-    "scriptID": "00003557",
-    "brandName": "Xalatan",
-    "indications": "",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "authCode": ""
-  },
-  {
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "authRxNumber": "00032185",
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "age": "",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "date": "2022-06-22",
-    "maxQuantity": "1",
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "authCode": "",
-    "scriptID": "00003556",
-    "repeats": "5",
-    "justification": "",
-    "maxRepeats": "5",
-    "brandName": "Xalatan",
-    "authRequired": false,
-    "pbsRx": true,
-    "indications": "",
-    "quantity": "1",
-    "verified": true,
-    "compounded": false,
-    "substitutePermitted": true
-  },
-  {
-    "date": "2022-06-22",
-    "brandName": "Xalatan",
-    "indications": "",
-    "authRxNumber": "00032174",
-    "authRequired": false,
-    "verified": true,
-    "dosage": "Once nightly OU",
-    "includeBrand": true,
-    "dateTime": "Jun 22, 2022, 10:15 AM",
-    "brandOnly": false,
-    "itemCode": "5552F",
-    "compounded": false,
-    "authCode": "",
-    "maxQuantity": "1",
-    "scriptID": "00003555",
-    "prevAuth": false,
-    "pbsRx": true,
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "quantity": "1",
-    "substitutePermitted": true,
-    "maxRepeats": "5",
-    "justification": "",
-    "age": "",
-    "repeats": "5"
-  },
-  {
-    "quantity": "1",
-    "repeats": "5",
-    "compounded": false,
-    "scriptID": "00003554",
-    "substitutePermitted": true,
-    "indications": "",
-    "itemCode": "5552F",
-    "prevAuth": false,
-    "brandName": "Xalatan",
-    "brandOnly": false,
-    "includeBrand": true,
-    "date": "2022-06-22",
-    "dateTime": "Jun 22, 2022, 10:14 AM",
-    "activeIngredient": "latanoprost 0.005% eye drops, 2.5 mL",
-    "maxRepeats": "5",
-    "pbsRx": true,
-    "justification": "",
-    "age": "",
-    "maxQuantity": "1",
-    "authRxNumber": "00032163",
-    "dosage": "Once nightly OU",
-    "authRequired": false,
-    "verified": true,
-    "authCode": ""
-  }
-]
-
-export default testScripts;
